Move context to the invocation
