+++ protocol/context.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package common
+package protocol
 
 import (
 	"context"
